refactor: mainElement 위치 수정
